Tests for qual args added.

+++ sh.d/test-small.sh
@@ -149,6 +149,7 @@ EXEC_ORDER=0
 TIMEOUT_CMD=("$(type -p timeout)" "-s" "TERM" "${TIMEOUT}"s)
 
 function AM_EXEC() {
+    EXEC_ORDER=$((EXEC_ORDER + 1))
     echo "EXEC ${EXEC_ORDER}: $(date '+%Y-%m-%d %H:%M:%S'): ${ART_CMD_ASSEMBLED[*]} $*"
     "${TIMEOUT_CMD[@]}" env \
         "ART_LOG_DIR=${OUT_DIR}/log_${EXEC_ORDER}.d" \
@@ -160,12 +161,33 @@ function AM_EXEC() {
         exit 1
     else
         echo "EXEC ${EXEC_ORDER}: $(date '+%Y-%m-%d %H:%M:%S'): Succeeded."
-        rm -f "${OUT_DIR}"/am_exec_"${EXEC_ORDER}".log
+        if [ -z "${KEEPLOG:-}" ]; then
+            rm -f "${OUT_DIR}"/am_exec_"${EXEC_ORDER}".log
+        fi
     fi
+}
+function AM_EXEC_SHOULD_FAIL() {
     EXEC_ORDER=$((EXEC_ORDER + 1))
-    return ${retval}
+    echo "EXEC ${EXEC_ORDER}: $(date '+%Y-%m-%d %H:%M:%S'): ${ART_CMD_ASSEMBLED[*]} $*"
+    set +e
+    "${TIMEOUT_CMD[@]}" env \
+        "ART_LOG_DIR=${OUT_DIR}/log_${EXEC_ORDER}.d" \
+        "${ART_CMD_ASSEMBLED[@]}" "$@" &>>"${OUT_DIR}"/am_exec_"${EXEC_ORDER}".log
+    retval=${?}
+    set -e
+    if [ ${retval} -eq 0 ]; then
+        echo "EXEC ${EXEC_ORDER}: $(date '+%Y-%m-%d %H:%M:%S'): Failed with exit code ${retval}" >&2
+        cat "${OUT_DIR}"/am_exec_"${EXEC_ORDER}".log >&2
+        exit 1
+    else
+        echo "EXEC ${EXEC_ORDER}: $(date '+%Y-%m-%d %H:%M:%S'): Succeeded."
+        if [ -z "${KEEPLOG:-}" ]; then
+            rm -f "${OUT_DIR}"/am_exec_"${EXEC_ORDER}".log
+        fi
+    fi
 }
 function APB_EXEC() {
+    EXEC_ORDER=$((EXEC_ORDER + 1))
     echo "EXEC ${EXEC_ORDER}: $(date '+%Y-%m-%d %H:%M:%S'): ${APB_CMD_ASSEMBLED[*]} $*"
     "${TIMEOUT_CMD[@]}" env \
         "ART_LOG_DIR=${OUT_DIR}/log_${EXEC_ORDER}.d" \
@@ -177,10 +199,10 @@ function APB_EXEC() {
         exit 1
     else
         echo "EXEC ${EXEC_ORDER}: $(date '+%Y-%m-%d %H:%M:%S'): Succeeded."
-        rm -f "${OUT_DIR}"/apb_exec_"${EXEC_ORDER}".log
+        if [ -z "${KEEPLOG:-}" ]; then
+            rm -f "${OUT_DIR}"/apb_exec_"${EXEC_ORDER}".log
+        fi
     fi
-    EXEC_ORDER=$((EXEC_ORDER + 1))
-    return ${retval}
 }
 
 rm -fr "${OUT_DIR}" # Remove previous runs
@@ -193,16 +215,17 @@ if [ "${HELP_VERSION_ONLY:-0}" == "1" ]; then
     rm -fr "${OUT_DIR}"
     exit 0
 fi
-. "${SHDIR}"/test-small.sh.d/00_out_fmts.sh       # Test all output is working
-. "${SHDIR}"/test-small.sh.d/01_fail.sh           # FASTA that would fail the simulator
-. "${SHDIR}"/test-small.sh.d/02_wgs.sh            # WGS mode (with constant coverage)
-. "${SHDIR}"/test-small.sh.d/03_trans_constcov.sh # Transcript mode with constant coverage
-. "${SHDIR}"/test-small.sh.d/04_tmpl_constcov.sh  # Template mode with constant coverage
-. "${SHDIR}"/test-small.sh.d/05_trans_scov.sh     # Transcript mode with stranded/strandless coverage
-. "${SHDIR}"/test-small.sh.d/06_tmpl_scov.sh      # Template mode with stranded/strandless coverage
-. "${SHDIR}"/test-small.sh.d/07_tmpl_pbsim3.sh    # Transcript mode with pbsim3-formatted coverage
-. "${SHDIR}"/test-small.sh.d/08_trans_pbsim3.sh   # Template mode with pbsim3-formatted coverage
-. "${SHDIR}"/test-small.sh.d/09_test_seedalloc.sh # Test seed allocation consistency
-. "${SHDIR}"/test-small.sh.d/21-apb-se.sh         # APB single-end test
-. "${SHDIR}"/test-small.sh.d/22-apb-pe.sh         # APB paired-end test
-rm -d "${OUT_DIR}"                                # Which should now be empty
+. "${SHDIR}"/test-small.sh.d/00_out_fmts.sh           # Test all output is working
+. "${SHDIR}"/test-small.sh.d/01_fail.sh               # FASTA that would fail the simulator
+. "${SHDIR}"/test-small.sh.d/02_wgs.sh                # WGS mode (with constant coverage)
+. "${SHDIR}"/test-small.sh.d/03_trans_constcov.sh     # Transcript mode with constant coverage
+. "${SHDIR}"/test-small.sh.d/04_tmpl_constcov.sh      # Template mode with constant coverage
+. "${SHDIR}"/test-small.sh.d/05_trans_scov.sh         # Transcript mode with stranded/strandless coverage
+. "${SHDIR}"/test-small.sh.d/06_tmpl_scov.sh          # Template mode with stranded/strandless coverage
+. "${SHDIR}"/test-small.sh.d/07_tmpl_pbsim3.sh        # Transcript mode with pbsim3-formatted coverage
+. "${SHDIR}"/test-small.sh.d/08_trans_pbsim3.sh       # Template mode with pbsim3-formatted coverage
+. "${SHDIR}"/test-small.sh.d/09_test_seedalloc.sh     # Test seed allocation consistency
+. "${SHDIR}"/test-small.sh.d/10_test_qual_file_arg.sh # Test quality profile arguments
+. "${SHDIR}"/test-small.sh.d/21-apb-se.sh             # APB single-end test
+. "${SHDIR}"/test-small.sh.d/22-apb-pe.sh             # APB paired-end test
+rm -d "${OUT_DIR}"                                    # Which should now be empty
